feat(backend): add Makefile for building backend services
This commit adds a new Makefile to the backend directory. The Makefile includes the necessary build configurations and variables for building the backend services. It also sets the NTP_DOCKER_IMAGE_NAME_PREFIX variable to the appropriate value.

The Makefile includes the following targets:
- BUILD_DIR: Specifies the build directory for the backend services.
- DM_SERVICE: Specifies the device manager service.
- SERVICES: Specifies the list of services to be built.
- CGO_ENABLED: Specifies whether CGO is enabled or not. The default value is 0.
- GOARCH: Specifies the target architecture for the build. The default value is amd64.
- VERSION: Specifies the version of the codebase. It is determined using the git describe command.
- COMMIT: Specifies the commit hash of the codebase. It is determined using the git rev-parse command.

These additions will make it easier to build the backend services and manage their dependencies.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>