feat(embedded/main): Add support for controlling LED strips in different locations
This commit adds support for controlling LED strips in different locations. It introduces three instances of the WS2812FX library, one for each location: Thika, Kiambu, and Juja. Each instance is initialized with the respective LED count and pin number. This allows for independent control of the LED strips in each location.

The code changes can be seen in the embedded/main/main.ino file.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>