docs(support): Add SUPPORT.md file
This commit adds a new SUPPORT.md file to provide guidelines and insights on where to find assistance with the nairobi-traffic-pcb projects. The file includes information on formulating questions, conducting searches, providing relevant code samples, and the importance of investing effort in framing questions effectively. It also includes a note to read the code of conduct before participating in the community.

The purpose of this commit is to improve the documentation and support resources for the project.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>