feat(README): Update description of how the PCB works
The README file has been updated to provide a more accurate and detailed description of how the PCB works. Specifically, the section explaining the use of cloud infrastructure to control the PCB in traffic mode has been added. This update aims to improve the understanding of the PCB functionality for potential users.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>